fix javascript, moved admin routes, moved helper functions
