+++ app/__init__.py
@@ -110,9 +110,11 @@ def create_app(config_class=Config):
     # Register blueprints
     from app.auth import bp as auth_bp
     from app.main import bp as main_bp
+    from app.main.admin_routes import admin_bp
 
     app.register_blueprint(main_bp)
     app.register_blueprint(auth_bp)
+    app.register_blueprint(admin_bp)
 
     @app.before_request
     def before_request_hook():
